task(core): fix burger, footer, navbar styles & some routing bugs
