+++ components/common/layout/CoreLayout.js
