+++ components/common/layout/PageLayout.js
@@ -3,7 +3,7 @@ import React from 'react';
 import Footer from './Footer';
 import Header from './Header';
 import CoreLayout from './CoreLayout';
-import LocaleContext from './LocaleContext';
+import LocaleContext from '../LocaleContext';
 
 export default ({ children, title, url }) => (
   <LocaleContext.Provider value={url.query.lang}>
