+++ constants/social.js
@@ -1,0 +1,28 @@
+export const NETWORKS = [
+  {
+    name: 'facebook',
+    link: 'https://facebook.com/wir_by',
+  },
+  {
+    name: 'vk',
+    link: 'https://vk.com/wir_by',
+  },
+  {
+    name: 'twitter',
+    link: 'https://twitter.com/wir_by',
+  },
+  {
+    name: 'telegram',
+    link: 'https://t.me/wir_by',
+  },
+  {
+    name: 'youtube',
+    link: 'https://www.youtube.com/channel/UCoj_6A55mEPeba8ZfBHqfCw',
+  },
+  {
+    name: 'instagram',
+    link: 'https://instagram.com/wir_by',
+  },
+];
+
+export const stub = 1;
