+++ package.json
@@ -32,7 +32,7 @@
     "@storybook/addon-info": "^3.2.19",
     "@storybook/addon-knobs": "^3.2.19",
     "@storybook/react": "^3.2.19",
-    "babajka-markup": "^0.6.0",
+    "babajka-markup": "^0.6.2",
     "classnames": "^2.2.5",
     "cookie-parser": "^1.4.3",
     "downshift": "^1.31.2",
